Change get_kv to not set the kv if missing
This prevents the last timestamp to be updated in the KV store at the start of the flow,
when we don't yet know if the flow execution will be successful.

It's unexpected to me that a function called 'get_kv' would perform a
write. If this is the desired behavior, then `setdefault` would be a
better name for the function. See:
https://docs.python.org/3/library/stdtypes.html#dict.setdefault