Braze Pocket Hits updates (#40)
* Initial Braze client

* Successfully sends request to Braze

* Send events and purchases to Braze

* Delete users and cleanup

* Remove premium purchase events

These events are not required for Pocket Hits, and once events are sent to Braze, they cannot be undone.
We should first define and signoff on the exact event that we need before we send it.
Event can be backfilled if necessary.

* Allow multiple triggers to be passed in

* Move chunks function into utils

* Create new user aliases

* Change get_kv to not set the kv if missing

This prevents the last timestamp to be updated in the KV store at the start of the flow,
when we don't yet know if the flow execution will be successful.

It's unexpected to me that a function called 'get_kv' would perform a
write. If this is the desired behavior, then `setdefault` would be a
better name for the function. See:
https://docs.python.org/3/library/stdtypes.html#dict.setdefault

* Remove None values and refactor email alias filters

* Events and attributes are successfully updated in Braze

* Add user identify functions

* Implement subscribing users

* Call identify users when emails are changed

* Make Snowflake schema configurable for local development

* Fix identify user error: alias objects must be a hash

* Implement Braze email validation

Based on https://www.braze.com/docs/user_guide/onboarding_with_braze/email_setup/email_validation/#email-validation

* Switch to simplified email validation

* Refactor Braze code

* Add mapping

* Add comments

* Schedule flow run

* Run Dbt orchestration on a schedule

* Add parameter for Snowflake table

* Don't alias or identify logged-in users

See https://pocket.slack.com/archives/C02ABJCHXQ9/p1648481237496609

For simplicity, keep the Pocket Hits signup page and logged-in users separate.

* Revert removing aliases for Pocket users

* Remove duplicate upstream_tasks definition

* Remove debug code

* Change Dbt orchestration interval to 10 minutes and fix dependencies

* Add parameter store parameters to ReadMe

* Set custom names for flow-of-flow

Currently they're all called create_flow_run or wait_for_flow_run

* Revert "Set custom names for flow-of-flow"

This reverts commit 23911bea8adfa7eb6c229ab027faded4fd74efee.